Automatically generate components/mod.rs
