+++ pgml-apps/cargo-pgml-components/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cargo-pgml-components"
-version = "0.1.5"
+version = "0.1.6"
 edition = "2021"
 authors = ["PostgresML <team@postgresml.org>"]
 license = "MIT"
