+++ pgml-apps/cargo-pgml-components/src/main.rs
@@ -4,7 +4,7 @@ use clap::{Args, Parser, Subcommand};
 use convert_case::{Case, Casing};
 use glob::glob;
 use std::env::{current_dir, set_current_dir};
-use std::fs::{create_dir_all, read_to_string, remove_file, File};
+use std::fs::{create_dir_all, read_to_string, remove_file, File, read_dir};
 use std::io::Write;
 use std::path::Path;
 use std::process::{exit, Command};
@@ -93,6 +93,8 @@ enum Commands {
         #[arg(short, long, default_value = "false")]
         overwrite: bool,
     },
+
+    UpdateComponents {},
 }
 
 fn main() {
@@ -103,6 +105,8 @@ fn main() {
         CargoSubcommands::PgmlComponents(pgml_commands) => match pgml_commands.command {
             Commands::Bundle {} => bundle(pgml_commands.project_path),
             Commands::AddComponent { name, overwrite } => add_component(name, overwrite),
+            Commands::UpdateComponents {} => update_components(),
+
         },
     }
 }
@@ -393,9 +397,32 @@ fn add_component(name: String, overwrite: bool) {
     let sass_file = File::create(sass_path).expect("failed to create sass file");
     drop(sass_file);
 
-    // let mut components_list = File::create("src/components/components.rs").expect("failed to create src/components/components.rs");
-    // let components = read_dir("src/components").expect("failed to read components directory");
-
     println!("Component '{}' created successfully", folder.display());
-    println!("Don't forget to add it to src/components/mod.rs");
+    update_components();
+}
+
+fn update_components() {
+    let mut file = File::create("src/components/mod.rs").expect("failed to create mod.rs");
+
+    writeln!(&mut file, "// This file is automatically generated by cargo-pgml-components.").expect("failed to write to mod.rs");
+    writeln!(&mut file, "// Do not modify it directly.").expect("failed to write to mod.rs");
+    writeln!(&mut file, "mod component;").expect("failed to write to mod.rs");
+    writeln!(&mut file, "pub(crate) use component::{{component, Component}};").expect("failed to write to mod.rs");
+
+    for component in read_dir("src/components").expect("failed to read components directory") {
+        let path = component.expect("dir entry").path();
+
+        if path.is_file() {
+            continue;
+        }
+
+        let components = path.components();
+        let component_name = components.clone().last().expect("component_name").as_os_str().to_str().unwrap();
+        let module = components.skip(2).map(|c| c.as_os_str().to_str().unwrap()).collect::<Vec<&str>>().join("::");
+        // let module = format!("crate::{}", module);
+        let component_name = component_name.to_case(Case::UpperCamel);
+
+        writeln!(&mut file, "pub mod {};", module).expect("failed to write to mod.rs");
+        writeln!(&mut file, "pub use {}::{};", module, component_name).expect("failed to write to mod.rs");
+    }
 }
