+++ pgml-dashboard/src/components/breadcrumbs/mod.rs
@@ -1,0 +1,17 @@
+use crate::components::component;
+use crate::components::NavLink;
+use sailfish::TemplateOnce;
+
+#[derive(TemplateOnce)]
+#[template(path = "breadcrumbs/template.html")]
+pub struct Breadcrumbs<'a> {
+    pub links: Vec<NavLink<'a>>,
+}
+
+impl<'a> Breadcrumbs<'a> {
+    pub fn render(links: Vec<NavLink<'a>>) -> String {
+        Breadcrumbs { links }.render_once().unwrap()
+    }
+}
+
+component!(Breadcrumbs, 'a);
