+++ pgml-dashboard/src/components/github_icon/template.html
