+++ pgml-dashboard/src/components/left_nav_menu/template.html
