+++ pgml-dashboard/src/components/left_nav_web_app/left_nav_web_app_controller.js
@@ -1,0 +1,14 @@
+import { Controller } from '@hotwired/stimulus'
+
+export default class extends Controller {
+  static targets = []
+  static outlets = []
+
+  initialize() {
+    console.log('Initialized left-nav-web-app')
+  }
+
+  connect() {}
+
+  disconnect() {}
+}
