+++ pgml-dashboard/src/components/navbar/mod.rs
@@ -1,0 +1,24 @@
+use crate::components::component;
+use crate::models;
+use crate::utils::config;
+use sailfish::TemplateOnce;
+
+#[derive(TemplateOnce)]
+#[template(path = "layout/nav/top.html")]
+pub struct Navbar {
+    pub current_user: Option<models::User>,
+    pub standalone_dashboard: bool,
+}
+
+impl Navbar {
+    pub fn render(user: Option<models::User>) -> String {
+        Navbar {
+            current_user: user,
+            standalone_dashboard: config::standalone_dashboard(),
+        }
+        .render_once()
+        .unwrap()
+    }
+}
+
+component!(Navbar);
