+++ pgml-dashboard/src/components/navbar/navbar_controller.js
@@ -1,0 +1,14 @@
+import { Controller } from '@hotwired/stimulus'
+
+export default class extends Controller {
+  static targets = []
+  static outlets = []
+
+  initialize() {
+    console.log('Initialized navbar')
+  }
+
+  connect() {}
+
+  disconnect() {}
+}
