+++ pgml-dashboard/src/components/navbar_web_app/template.html
@@ -18,7 +18,7 @@
   
         <!-- Button to toggle collapsed menu for less than lg screens -->
         <button class="navbar-toggler collapsed topnav-controlls" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-          <% include!("../../components/profile_icon.html"); %>
+          <% include!("../../../templates/components/profile_icon.html"); %>
         </button>
       </div>
 
@@ -59,7 +59,7 @@
             <% if !account_management_nav.links.is_empty() { %>
             <li class="d-none d-lg-flex nav-item align-items-center dropdown">
               <a class="dropdown-toggle topnav-controlls" href="#" role="button" data-bs-toggle="dropdown" data-bs-offset="10, 15" aria-expanded="false">
-                <% include!("../../components/profile_icon.html"); %>
+                <% include!("../../../templates/components/profile_icon.html"); %>
               </a>
               <!-- Account management for lg and up screens -->
               <ul class="dropdown-menu dropdown-menu-end">
@@ -158,4 +158,4 @@
     </nav>
   </div>
 
-  <% include!("../../components/search_modal.html");%>
+  <% include!("../../../templates/components/search_modal.html");%>
