+++ pgml-dashboard/src/components/postgres_logo/postgres_logo.scss
@@ -1,0 +1,6 @@
+.postgres-logo {
+    display: flex; 
+    align-items: center;
+    gap: calc($spacer / 2);
+    font-size: 24px;
+}
