+++ pgml-dashboard/src/components/postgres_logo/template.html
