+++ pgml-dashboard/src/components/static_nav/static_nav_controller.js
@@ -1,0 +1,14 @@
+import { Controller } from '@hotwired/stimulus'
+
+export default class extends Controller {
+  static targets = []
+  static outlets = []
+
+  initialize() {
+    console.log('Initialized static-nav')
+  }
+
+  connect() {}
+
+  disconnect() {}
+}
