+++ pgml-dashboard/src/components/static_nav/template.html
@@ -1,0 +1,3 @@
+<div data-controller="static-nav">
+  <%= value %>
+</div>
