+++ pgml-dashboard/static/css/modules.scss
@@ -1,2 +1,7 @@
+@import "../../src/components/left_nav_menu/left_nav_menu.scss";
+@import "../../src/components/left_nav_web_app/left_nav_web_app.scss";
 @import "../../src/components/modal/modal.scss";
+@import "../../src/components/navbar/navbar.scss";
+@import "../../src/components/postgres_logo/postgres_logo.scss";
+@import "../../src/components/static_nav/static_nav.scss";
 @import "../../src/components/test_component/test_component.scss";
