+++ pgml-dashboard/static/css/scss/components/_buttons.scss
@@ -244,13 +244,6 @@
     font-weight: $font-weight-medium;
 }
 
-.postgres-logo {
-    display: flex; 
-    align-items: center;
-    gap: calc($spacer / 2);
-    font-size: 24px;
-}
-
 .btn-dropdown {
     @extend .btn;
     border-radius: $border-radius;
