--- pgml-dashboard/templates/components/box.html
@@ -1,8 +1,0 @@
-<div class="flex-grow-1">
-  <div class="card rounded-1">
-    <div class="card-body d-flex flex-column justify-content-center align-items-center p-3">
-      <h6 class="h6"><%= name %></h6>
-      <p class="fs-6 font-weight-bold mb-0"><%- value %></p>
-    </div>
-  </div>
-</div>
